pyproject.toml
[project]
name = "hse-mlops-project"
version = "0.0.1"
description = "AI generated text detection. HSE MLOps project"
authors = [
    {name = "vakhibin@hse.edu.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11.13,<3.15"
dependencies = [
    "torch (>=2.9.1,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.4.0,<3.0.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "pytorch-lightning (>=2.6.0,<3.0.0)",
    "torchmetrics (>=1.8.2,<2.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "kagglehub (>=0.3.13,<0.4.0)",
    "fire (>=0.7.1,<0.8.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "lightning (>=2.6.0,<3.0.0)",
    "dvc (>=3.65.0,<4.0.0)",
    "tensorboard (>=2.20.0,<3.0.0)",
    "tensorboardx (>=2.6.4,<3.0.0)",
    "mlflow (>=3.8.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
